move: 🚚 Move Library specific encode-decode functions to a separate new file `libraries.ts`
